backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.0
dj-database-url==2.1.0
reportlab==4.0.9
